+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ Release notes
 ### Unreleased
 
 * Fix: Make strict mode work with query parameters parsed using request parser's `ParseAsync`/`ParseTask`
+* Added `SKipLink` to custom operations, which will make Felicity not add the link to the resource's `links` object. Using this is [required in order to be JSON:API compliant](https://github.com/json-api/json-api/issues/1656), but it is not enabled by default for backward compatibility reasons.
 
 ### 0.20.10 (2022-09-16)